Move 'saveAIResponse' into the file 'aiPrompt.tsx'
