+++ src/components/ai/aiPrompt.tsx
@@ -1,6 +1,7 @@
 "use client";
 
 import { model } from "@/lib/firebase";
+import { doc, getDoc, updateDoc } from "firebase/firestore";
 
 export const AIPrompt = `Please take this text corpus submitted by a user and parse the following information from it:
 
@@ -64,4 +65,66 @@ export async function getAIResponse(prompt: string, corpus: string) {
     const response = result.response;
     const text = response.text();
     return text;
+}
+
+export async function saveAIResponse(responseObj: any, user: any, db: any) {
+    if (user) {
+        const documentRef = doc(db, "users", user.uid);
+        try {
+            const document = await getDoc(documentRef);
+            if (!document.exists()) {
+                return;
+            }
+            // Extract full name and save to userProfile
+            try {
+                await updateDoc(documentRef, { "resumeFields.fullName": responseObj.fullName });
+            } catch (error) {
+                ;
+            }
+            // Extract summary and save to userProfile
+            try {
+                await updateDoc(documentRef, { "resumeFields.summary": responseObj.summary });
+            } catch (error) {
+                ;
+            }
+            // Extract email and save to userProfile
+            try {
+                await updateDoc(documentRef, { "resumeFields.contact.email": responseObj.contact.email });
+            } catch (error) {
+                ;
+            }
+            // Extract location and save to userProfile
+            try {
+                await updateDoc(documentRef, { "resumeFields.contact.location": responseObj.contact.location });
+            } catch (error) {
+                ;
+            }
+            // Extract phone and save to userProfile
+            try {
+                await updateDoc(documentRef, { "resumeFields.contact.phone": responseObj.contact.phone });
+            } catch (error) {
+                ;
+            }
+            // Extract list of skills and save to userProfile
+            try {
+                await updateDoc(documentRef, { "resumeFields.skills": responseObj.skills });
+            } catch (error) {
+                ;
+            }
+            // Extract list of work experiences and save to userProfile
+            try {
+                await updateDoc(documentRef, { "resumeFields.workExperience": responseObj.workExperience });
+            } catch (error) {
+                ;
+            }
+            // Extract list of education credentials and save to userProfile
+            try {
+                await updateDoc(documentRef, { "resumeFields.education": responseObj.education });
+            } catch (error) {
+                ;
+            }
+        } catch (error) {
+            console.error("Error: could not retrieve document;", error);
+        }
+    }
 }
